centralize logic to update the root readme.md
